requirements.txt
# gradio~=5.13.1
# uvicorn~=0.34.0
# fastapi~=0.115.7
# pydantic~=2.10.6
# fsspec~=2024.12.0
# torch~=2.5.1
# numpy~=1.26.4
# transformers~=4.48.1
# tqdm~=4.67.1
# pandas~=2.2.3
fastapi==0.115.7
pydantic==2.10.6
langchain==0.3.15
langchain-groq==0.2.3
pandas==2.2.3
uvicorn==0.34.0 
python-dotenv==1.0.1
langchain-community==0.3.15
langgraph~=0.2.68
langchain-mistralai~=0.2.5